# MINOR: reduce allocations in log start and recovery checkpoints