--- a/core/src/main/scala/kafka/utils/Pool.scala
+++ b/core/src/main/scala/kafka/utils/Pool.scala
@@ -18,7 +18,6 @@
 package kafka.utils
 
 import java.util.concurrent._
-import java.util.function.BiConsumer
 
 import org.apache.kafka.common.KafkaException