@lbradstreet 2020-04-19
Updated benchmarks with AnyRefMap change:
```
Benchmark                                         (numPartitions)  (numTopics)   Mode  Cnt  Score   Error   Units
CheckpointBench.measureCheckpointLogStartOffsets                3          100  thrpt   15  2.673 ± 0.027  ops/ms
CheckpointBench.measureCheckpointLogStartOffsets                3         1000  thrpt   15  1.066 ± 0.012  ops/ms
CheckpointBench.measureCheckpointLogStartOffsets                3         2000  thrpt   15  0.609 ± 0.007  ops/ms
```

Looks marginally better. I'm re-running with `-prof gc` so I can include the final results in the commit.