@lbradstreet 2020-04-24
I have updated the benchmark results in the OP with the new `forEach` code. It is strictly better than the old `foreach` code.